move sharing related tests to new context
